// src/types/collection.ts
export type TCollection = {
  id: number;
  store_id: number;
  name: string;
  name_eng: string;
  description: string;
  description_eng: string;
  seo: string;
  seodescription: string;
  seokeyword: string;
  link: string;
  banner_url: string;
  brand_id: number;
  type: number;
  position: number;
};